Removes space from in front of `:` when using `:location` option
When utilizing the location option the log message would format as
`[debug] Module.function/airity : message`. After this it will format as
`[debug] Module.function/airity: message`. Also updates the Readme to
point to where the options are defined.

The reason for `location` as an option comes down to simplicity and
format. The module and function are allowed as part fore the metadata,
but the output is much harder to read, IMO.

`module=Module function=function/airity [debug] Hello`

I feel this is harder because it doesn't match the documentation format
of module, function, and airity. It also moves the level further into
the message. I think this hides the data and increases cognitive
overhead.

Amos King @adkron <amos@binarynoggin.com>