Refactor tests to handshake that logs made it
This simplifies handling of the race condition where you log a message
and then expect RingLogger to have it. Since logging is asynchronous,
this isn't guaranteed. To make this work, the test has to
RingLogger.attach and then the handshake is the message being "printed".
This isn't appropriate for all tests especially those that test that
RingLogger.attach works.