Print in the context of the caller
This fixes the GenServer timeouts when next, tail, and grep have a lot
to print over a slow connection. Now if anything goes wrong or takes
unexpectantly long, it will come from an IO.binwrite call in the
caller's context. This had two unintended benefits:

1. One call to IO.binwrite. Apparently I missed that I had iodata
   before. This removes an Enum.each.
2. grep now grep's the entire log message instead of just the message
   part. You can grep on timestamps and debug level now.

The downside to #1 is that it had an effect on the unit tests since all
output comes at once. I updated the unit tests to expect this.