Switch to custom circular buffer implementation
This builds off in progress updates to the `circular_buffer` library to
optimize it for the circular buffer case. This update reduces the work
involved with storing log records. Currently other parts of RingLogger
dominate in timings so the improvement from this update can only be seen
if carefully monitoring the amount of memory used.