Fix setting colors enabled
The default for outputting color is determined by `IO.ANSI.enabled?`, but it
is applied in the `%RingLogger.Client.State{}` struct which is evaluated at
compile time so it will always be `false` and remove the color output.

This change forces in a `:colors` option key when building defaults if it
was not already present (which is normally the case). This will then cause
the config check to go through the configuring colors option which evaluates
`IO.ANSI.enabled?` at runtime for more accurate settings for the prints.